Rename CustomMessageControlPanel and change to small description
