+++ src/CaptainsLog/CaptainsLogModPlugin.java
@@ -1,6 +1,6 @@
 package CaptainsLog;
 
-import CaptainsLog.campaign.intel.FleetLogIntel;
+import CaptainsLog.campaign.intel.CustomMessageControlPanel;
 import CaptainsLog.campaign.listeners.LocationChangeListener;
 import CaptainsLog.campaign.listeners.RemovableSalvageListener;
 import CaptainsLog.campaign.listeners.SettingsChangeListener;
@@ -15,7 +15,7 @@ public class CaptainsLogModPlugin extends BaseModPlugin {
 
     @Override
     public void onGameLoad(boolean newGame) {
-        FleetLogIntel.getInstance();
+        CustomMessageControlPanel.getInstance();
 
         Global.getSector().addTransientScript(new CaptainsLogEveryFrame());
         Global.getSector().addTransientScript(new RuinObserver());
